Detailed list of semiconductors added as commodities
